vulkan: read the semi-transparency blend destination at highp
The feedback subpass input in primitive_feedback.frag - the framebuffer
destination read for PSX semi-transparency (B+F / avg / B-F / B+F/4) - was
declared mediump. With the scaled framebuffer now R16G16B16A16_SFLOAT,
that destination carries both the sub-8-bit blend precision and the >1.0
additive accumulation, and a GPU that honours mediump as fp16 would
truncate the read. Declare it highp so the 16F precision survives the
blend.

No effect on the desktop target (mediump is fp32 there, so the result is
identical - only the RelaxedPrecision decoration differs); this is a
correctness fix for the 16F path and lower-end Vulkan devices. Shared by
the SDR pipelines, where it is likewise a no-op on fp32.

+++ rhi/shaders_vulkan/primitive_feedback.frag
@@ -6,9 +6,9 @@ precision highp int;
 #include "primitive.h"
 
 #ifdef MSAA
-layout(set = 0, binding = 3, input_attachment_index = 0) uniform mediump subpassInputMS uFeedbackFramebuffer;
+layout(set = 0, binding = 3, input_attachment_index = 0) uniform highp subpassInputMS uFeedbackFramebuffer;
 #else
-layout(set = 0, binding = 3, input_attachment_index = 0) uniform mediump subpassInput uFeedbackFramebuffer;
+layout(set = 0, binding = 3, input_attachment_index = 0) uniform highp subpassInput uFeedbackFramebuffer;
 #endif
 
 const int BLEND_ADD = 0;
